Rename st_gaudi_transformer_tokenize.py to st_gaudi_transformer.py
